+++ mscmcldap/api/spl/models.py
@@ -1,0 +1,136 @@
+# -*- coding: utf-8 -*-
+
+from __future__ import unicode_literals
+from django.db import models
+from django.utils.encoding import python_2_unicode_compatible
+
+
+#---------------------------------------------------------------------------------------------
+# Model para a view V_SPL_REUNIAO_COMISSAO
+#---------------------------------------------------------------------------------------------
+@python_2_unicode_compatible
+class v_spl_reuniao_comissao(models.Model):
+	class Meta:
+		managed = False
+		db_table = "v_spl_reuniao_comissao"
+
+	rec_id = models.IntegerField(primary_key=True)
+	con_id = models.IntegerField()
+	rec_tipo_reuniao = models.CharField(max_length=500)
+	rec_numero = models.CharField(max_length=100)
+	versao = models.IntegerField()
+	rec_data = models.DateField()
+
+	def __unicode__(self):
+		return self.rec_tipo_reuniao
+
+	def __str__(self):
+		return self.rec_tipo_reuniao
+
+
+#---------------------------------------------------------------------------------------------
+# Model para a view V_CONJUNTO_VEREADORES
+#---------------------------------------------------------------------------------------------
+@python_2_unicode_compatible
+class v_spl_conjunto_vereadores(models.Model):
+	class Meta:
+		managed = False
+		db_table = "v_spl_conjunto_vereadores"
+
+	con_id = models.IntegerField(primary_key=True)
+	con_sigla = models.CharField(max_length=200)
+	con_maximo = models.IntegerField(blank=True, null=True)
+	ut_id = models.IntegerField()
+	ini_nome = models.CharField(max_length=500)
+	ini_ativa = models.BooleanField()
+	tin_id = models.IntegerField()
+	versao = models.IntegerField()
+	ini_codigo_prefeitura = models.IntegerField()
+	con_descricao = models.CharField(max_length=500)
+	ini_nome = models.CharField(max_length=500)
+	con_site_ordem = models.IntegerField()
+	con_email = models.CharField(max_length=500)
+
+	def __unicode__(self):
+		return self.ini_nome
+
+	def __str__(self):
+		return self.ini_nome
+
+
+#---------------------------------------------------------------------------------------------
+# Model para a view V_PAUTA_COMISSAO
+#---------------------------------------------------------------------------------------------
+@python_2_unicode_compatible
+class v_spl_pauta_comissao(models.Model):
+	class Meta:
+		managed = False
+		db_table = "v_spl_pauta_comissao"
+
+	pac_id = models.IntegerField(primary_key=True)
+	rec_id = models.IntegerField()
+	pac_liberada = models.BooleanField()
+	pac_notificada = models.BooleanField()
+	versao = models.IntegerField()
+
+	def __unicode__(self):
+		return self.pac_id
+
+	def __str__(self):
+		return self.pac_id
+
+
+# ---------------------------------------------------------------------------------------------
+# Model para a view V_VEREADOR
+# ---------------------------------------------------------------------------------------------
+@python_2_unicode_compatible
+class v_spl_vereador(models.Model):
+    class Meta:
+        managed = False
+        db_table = "v_spl_vereador"
+
+    ver_id = models.IntegerField(primary_key=True)
+    matricula = models.CharField(max_length=4)
+    ver_sexo = models.CharField(max_length=1)
+    ut_id = models.IntegerField()
+    ini_nome = models.CharField(max_length=100)
+    ini_ativa = models.BooleanField()
+    ver_nome_completo = models.CharField(max_length=100)
+    ini_codigo_prefeitura = models.IntegerField()
+    ver_site = models.CharField(max_length=250)
+    ver_biografia = models.TextField()
+    ver_redes_sociais = models.CharField(max_length=250)
+    ver_fone_principal = models.CharField(max_length=20)
+    ver_fones = models.CharField(max_length=100)
+    ver_legislaturas = models.CharField(max_length=100)
+    ver_localizacao = models.CharField(max_length=50)
+    ver_partido = models.CharField(max_length=100)
+    arq_id = models.IntegerField()
+    arq_id_biografia = models.IntegerField()
+
+    def __unicode__(self):
+        return self.ini_nome
+
+    def __str__(self):
+        return self.ini_nome
+
+
+# ---------------------------------------------------------------------------------------------
+# Model para a view V_CARGO_MESA
+# ---------------------------------------------------------------------------------------------
+@python_2_unicode_compatible
+class v_spl_cargos_mesa(models.Model):
+    class Meta:
+        managed = False
+        db_table = "v_spl_cargos_mesa"
+
+    matricula = models.CharField(primary_key=True, max_length=4)
+    ini_nome = models.TextField()
+    crg_nome = models.CharField(max_length=50)
+    crg_ordem = models.IntegerField()
+
+    def __unicode__(self):
+        return self.ini_nome
+
+    def __str__(self):
+        return self.ini_nome
